feat: implement Databricks-based data diff classes and connection configuration
