+++ ingestion/src/metadata/data_quality/validations/runtime_param_setter/base_diff_params_setter.py
@@ -140,6 +140,26 @@ class BaseTableParameter:
 
         try:
             connection_class = service_spec_patch.get_connection_class()
+            # Check if the data diff class is not BaseTableParameter, but an overridden table parameter, then get that
+            data_diff_class = service_spec_patch.get_data_diff_class()
+            # Check if method is truly overridden (not just different class)
+            if data_diff_class and data_diff_class is not BaseTableParameter:
+                # Compare method objects to ensure it's overridden
+                class_method = getattr(
+                    data_diff_class, "_get_service_connection_config", None
+                )
+                base_method = getattr(
+                    BaseTableParameter, "_get_service_connection_config", None
+                )
+
+                if class_method is not base_method:
+                    # Method is overridden, create instance and call it
+                    instance = data_diff_class()
+                    # Call as instance method to avoid recursion
+                    return instance._get_service_connection_config.__func__(
+                        instance, service_connection_config
+                    )
+
             if not connection_class:
                 return (
                     str(get_connection(service_connection_config).url)
@@ -188,7 +208,10 @@ class BaseTableParameter:
         )
         # path needs to include the database AND schema in some of the connectors
         if hasattr(db_service.connection.config, "supportsDatabase"):
-            kwargs["path"] = f"/{database}"
+            if kwargs["scheme"] in {Dialects.UnityCatalog, Dialects.Databricks}:
+                kwargs["query"] = f"catalog={database}"
+            else:
+                kwargs["path"] = f"/{database}"
         if kwargs["scheme"] in {Dialects.MSSQL, Dialects.Snowflake, Dialects.Trino}:
             kwargs["path"] = f"/{database}/{schema}"
         return url._replace(**kwargs).geturl()
