+++ ingestion/src/metadata/ingestion/source/database/databricks/data_diff/__init__.py
@@ -1,0 +1,1 @@
+"""Unity Catalog data diff module"""
